feat: 팔로잉 목록 페이지 url 변경 및 로딩 UI 추가
